requirements.txt
anyio==3.5.0
asgiref==3.5.0
backports.cached-property==1.0.1
click==8.1.2
graphql-core==3.2.0
h11==0.13.0
idna==3.3
mercurial==5.6.1
Pygments==2.11.2
python-dateutil==2.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
strawberry-graphql==0.105.0
typing-extensions==4.1.1
uvicorn==0.17.6
